scripts/requirements.txt
watchdog
supabase
python-pptx
python-docx
pypdf
Pillow
requests
python-dotenv

